+++ src/services/requests.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-export default class Api {
+export default class Requests {
 
     static buildParams(path, body) {
         path += "?";
@@ -14,8 +14,7 @@ export default class Api {
     static get(path, body) {
         if (body)
             path = process.env.REACT_APP_API_URL + this.buildParams(path, body);
-        console.log("REACT_APP_API_URL: ",process.env.REACT_APP_API_URL);
-        return axios.get(path)
+        return axios.get(path, {withCredentials : true})
             .then(res => {
                 return res;
             })
